light-desk: move audio to shared assets directory
