+++ light-desk/src/frontend/util/audio.ts
@@ -1,11 +1,6 @@
 import { AudioFile, AUDIO_FILES } from '../../shared/static';
 
 /**
- * The root URL used for this light desk
- */
-const ROOT_URL = window.location.pathname;
-
-/**
  * A fixed-number collection of Audio elements for a single sound file, to allow for multi-playback of a single sound file
  */
 class AudioPool {
@@ -16,7 +11,9 @@ class AudioPool {
   constructor(sample: AudioFile) {
     this.startTime = sample.startTime;
     for (let i = 0; i < sample.poolSize; i++) {
-      const audio = new Audio(`${ROOT_URL}audio/${sample.file}`);
+      const audio = new Audio(
+        new URL(`audio/${sample.file}`, window.location.href).href
+      );
       audio.volume = sample.volume;
       this.pool.push(audio);
     }
